Generic Exception Handler Project Library for .NET Framework 4.6.1 VS 2017
This is a generic Exception Handler Library with a provision to log the exception in advanced message format in various medias either in TXT / Excel / CSV or Database. It has a provision to Log exception with either success or failure message status and message object to send at client side along with final response which can help out to process UI based on final status.